fix(cli): config cache invalidation + refactors (#5500)
