+++ packages/cli/src/command.js
@@ -3,7 +3,8 @@ import path from 'path'
 import consola from 'consola'
 import minimist from 'minimist'
 import { name, version } from '../package.json'
-import { loadNuxtConfig, forceExit } from './utils'
+import { forceExit } from './utils'
+import { loadNuxtConfig } from './utils/config'
 import { indent, foldLines, colorize } from './utils/formatting'
 import { startSpaces, optionSpaces, forceExitTimeout } from './utils/constants'
 import { detectTypeScript } from './utils/typescript'
