+++ packages/cli/src/index.js
@@ -7,4 +7,4 @@ export const imports = _imports
 export { default as NuxtCommand } from './command'
 export { default as setup } from './setup'
 export { default as run } from './run'
-export { loadNuxtConfig } from './utils'
+export { loadNuxtConfig } from './utils/config'
