+++ packages/cli/src/utils/index.js
@@ -1,125 +1,15 @@
 import path from 'path'
-import consola from 'consola'
-import esm from 'esm'
 import exit from 'exit'
-import defaultsDeep from 'lodash/defaultsDeep'
-import { defaultNuxtConfigFile, getDefaultNuxtConfig } from '@nuxt/config'
+
 import { lock } from '@nuxt/utils'
 import chalk from 'chalk'
-import prettyBytes from 'pretty-bytes'
 import env from 'std-env'
-import { successBox, warningBox } from './formatting'
-
-const esmOptions = {
-  cache: false,
-  cjs: {
-    cache: true,
-    vars: true,
-    namedExports: true
-  }
-}
+import { warningBox } from './formatting'
 
 export const eventsMapping = {
   add: { icon: '+', color: 'green', action: 'Created' },
   change: { icon: env.windows ? '»' : '↻', color: 'blue', action: 'Updated' },
   unlink: { icon: '-', color: 'red', action: 'Removed' }
-}
-
-export async function loadNuxtConfig(argv) {
-  const rootDir = path.resolve(argv._[0] || '.')
-  let nuxtConfigFile
-  let options = {}
-
-  try {
-    nuxtConfigFile = require.resolve(path.resolve(rootDir, argv['config-file']))
-  } catch (e) {
-    if (e.code !== 'MODULE_NOT_FOUND') {
-      throw (e)
-    } else if (argv['config-file'] !== defaultNuxtConfigFile) {
-      consola.fatal('Could not load config file: ' + argv['config-file'])
-    }
-  }
-
-  if (nuxtConfigFile) {
-    options = (nuxtConfigFile.endsWith('.ts') ? require(nuxtConfigFile) : esm(module, esmOptions)(nuxtConfigFile)) || {}
-    if (options.default) {
-      options = options.default
-    }
-
-    if (typeof options === 'function') {
-      try {
-        options = await options()
-        if (options.default) {
-          options = options.default
-        }
-      } catch (error) {
-        consola.error(error)
-        consola.fatal('Error while fetching async configuration')
-      }
-    }
-
-    // Keep _nuxtConfigFile for watching
-    options._nuxtConfigFile = nuxtConfigFile
-  }
-
-  if (typeof options.rootDir !== 'string') {
-    options.rootDir = rootDir
-  }
-
-  // Nuxt Mode
-  options.mode =
-    (argv.spa && 'spa') || (argv.universal && 'universal') || options.mode
-
-  // Server options
-  options.server = defaultsDeep({
-    port: argv.port || undefined,
-    host: argv.hostname || undefined,
-    socket: argv['unix-socket'] || undefined
-  }, options.server || {}, getDefaultNuxtConfig().server)
-
-  return options
-}
-
-export function showBanner(nuxt) {
-  if (env.test) {
-    return
-  }
-
-  if (env.minimalCLI) {
-    for (const listener of nuxt.server.listeners) {
-      consola.info('Listening on: ' + listener.url)
-    }
-    return
-  }
-
-  const titleLines = []
-  const messageLines = []
-
-  // Name and version
-  titleLines.push(`${chalk.green.bold('Nuxt.js')} ${nuxt.constructor.version}`)
-
-  // Running mode
-  titleLines.push(`Running in ${nuxt.options.dev ? chalk.bold.blue('development') : chalk.bold.green('production')} mode (${chalk.bold(nuxt.options.mode)})`)
-
-  if (nuxt.options._typescript && nuxt.options._typescript.runtime) {
-    titleLines.push(`TypeScript support is ${chalk.green.bold('enabled')}`)
-  }
-
-  // https://nodejs.org/api/process.html#process_process_memoryusage
-  const { heapUsed, rss } = process.memoryUsage()
-  titleLines.push(`Memory usage: ${chalk.bold(prettyBytes(heapUsed))} (RSS: ${prettyBytes(rss)})`)
-
-  // Listeners
-  for (const listener of nuxt.server.listeners) {
-    messageLines.push(chalk.bold('Listening on: ') + chalk.underline.blue(listener.url))
-  }
-
-  // Add custom badge messages
-  if (nuxt.options.cli.badgeMessages.length) {
-    messageLines.push('', ...nuxt.options.cli.badgeMessages)
-  }
-
-  process.stdout.write(successBox(messageLines.join('\n'), titleLines.join('\n')))
 }
 
 export function formatPath(filePath) {
